Replace SatangPro with Bitkub for THB quotes.
