+++ README.md
@@ -17,7 +17,7 @@ Available APIs:
 - Bit2C
 - BitMEX
 - BitPay
-- BitPay
+- Bitkub
 - Bitfinex
 - Bitso
 - Bitstamp
@@ -38,7 +38,6 @@ Available APIs:
 - MEXC
 - Nobitex
 - Paymium
-- Satang.pro
 - TomoX(TomoChain)
 - VccExchange(Vietnam)
 
@@ -116,7 +115,7 @@ On most systems you can use `journalctl /usr/bin/gnome-shell -f` to get log outp
 - https://github.com/volandku - BitMEX support
 - https://github.com/HawtDogFlvrWtr - CryptoCompare support
 - https://github.com/jpereira - Blinktrade support
-- https://github.com/rossigee - BX.in.th and Satang.pro support
+- https://github.com/rossigee - Bitkub support
 - https://github.com/WeirdFishBK201 - Huobi support
 - https://github.com/thanhnguyennguyen - HitBTC, Kucoin, VccExchange, BinanceFutures, FTX and Tomo support
 - https://github.com/luissantos - Bittrex support
